pyproject.toml
[project]
name = "MaaStarResonance"
version = "0.1.0"
description = "基于MAAFW的星痕共鸣黑盒测试工具"
authors = [
    {name = "233Official",email = "ayusummer233@vip.qq.com"}
]
license = {text = "LGPL-3.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "maafw>=4.5.5",
    "pre-commit>=4.3.0",
    "toml>=0.10.2",
    "tzdata>=2025.2",
]

[tool.setuptools]
packages = ["agent"]